+++ CHANGES.rst
@@ -5,6 +5,7 @@
 * New feature: Labeled enumerations.
 * Enhancement: load_models allows choice of permissions and takes additional
   permissions.
+* Rewrote requestargs view decorator for efficiency and ease of use
 
 0.3.11
 ------